% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast.R
\name{contrast_estimation_vg.unconstrained}
\alias{contrast_estimation_vg.unconstrained}
\title{Minimal contrast for unconstrained estimation with thinned varGamma}
\usage{
contrast_estimation_vg.unconstrained(
  params_init,
  rho_hat,
  K_hat,
  repulsionRange_hat,
  nSims,
  win,
  union_est = FALSE,
  w_rho = 1,
  q_rho = 1,
  p_rho = 2,
  w_K = 2,
  q_K = 1/4,
  p_K = 2
)
}
\arguments{
\item{params_init}{Initial guess for parameter vector}

\item{rho_hat}{Intensity estimated from data}

\item{K_hat}{K-function estimated from data}

\item{repulsionRange_hat}{Set to FALSE if the repulsion range parameter is in the params vector}

\item{nSims}{How many simulations to estimate K_hat}

\item{win}{Simulation window}

\item{union_est}{Set to TRUE in order to use union approach for estimation of K-function}

\item{w_rho}{Weight of intensity}

\item{q_rho}{Power to use inside parantheses in contrast for intensity}

\item{p_rho}{Power to use outside parantheses in contrast for intensity}

\item{w_K}{Weight of K-function}

\item{q_K}{Power to use inside parantheses in contrast for K-function}

\item{p_K}{Power to use outside parantheses in contrast for K-function}
}
\description{
Use minimal contrast estimation together with Monte Carlo calculations for K_theo for parameter estimation for Matern thinned VarGamma
}
