% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{contrast_function_importance_sampling}
\alias{contrast_function_importance_sampling}
\title{Use importance sampling to evaluate the contrast function}
\usage{
contrast_function_importance_sampling(
  patternSim,
  K_hat,
  rho_hat,
  wqp = c(1, 1, 1, 1, 1/4, 2)
)
}
\arguments{
\item{patternSim}{List of point patterns to use for estimation}

\item{K_hat}{estimated K-function for data}

\item{rho_hat}{Estimated intensity for data}

\item{wqp}{Weights for contrast function}
}
\description{
A short description...
}
