% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_SR1.R
\name{contrast_hessian_importance_sampling}
\alias{contrast_hessian_importance_sampling}
\title{Estimate Hessian at some parameter value using importance sampling}
\usage{
contrast_hessian_importance_sampling(
  params,
  params_sim,
  par_free_index,
  epsilon,
  simStepRes,
  normalized,
  rho_hat,
  K_hat,
  wq,
  f_val
)
}
\arguments{
\item{params}{Parameter values at which to aproximate the partial derivative}

\item{params_sim}{Parameter values used to simulate \code{simStepRes}}

\item{par_free_index}{Indices of parameters to estimate}

\item{epsilon}{Step for derivative calculation}

\item{simStepRes}{Result from the simulation step, see \code{\link{simulation_step}}. It must be the ensemble
simulated at \code{params_sim}.}

\item{normalized}{Normalize IS weights?}

\item{rho_hat}{Estimated intensity for data}

\item{K_hat}{Estimated K-function for data}

\item{wq}{Weights for contrast function}

\item{f_val}{Value of function at evaluation location}
}
\description{
See equation 8.21 in Nocedal and Wright.
}
