% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast.R
\name{contrast_theo_est}
\alias{contrast_theo_est}
\title{Contrast function}
\usage{
contrast_theo_est(params, K_est, K_theo, r_vec, dr, win, q = 1/4, p = 2)
}
\arguments{
\item{params}{Parameter vector(to be optimized)}

\item{K_est}{K-function estimated from data}

\item{K_theo}{Function for how to calculate K_theo}

\item{r_vec}{Vector with radii to use for contrast calculation}

\item{dr}{Step size for r_vec}

\item{win}{Simulation window}

\item{q}{Power to use in contrast(inner)}

\item{p}{Power to use in contrast(outer)}
}
\description{
Calculate contrast given params and an estimated K-function from data
}
