% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_SR1.R
\name{evaluate_contrast_and_gradient}
\alias{evaluate_contrast_and_gradient}
\title{Estimate both the contrast function and gradient of the contrast function using importance sampling}
\usage{
evaluate_contrast_and_gradient(
  simStepRes,
  params,
  params_0,
  rho_hat,
  K_hat,
  par_free_index,
  epsilon,
  normalized,
  wq
)
}
\arguments{
\item{simStepRes}{Result from the simulation step, see \code{\link{simulation_step}}.}

\item{params}{Parameters to use in estimation}

\item{params_0}{Parameters used for simulation}

\item{rho_hat}{Estimated intensity for data}

\item{K_hat}{Estimated K-function for data}

\item{par_free_index}{Indices of parameters to estimate}

\item{epsilon}{Step for derivative calculation}

\item{normalized}{Normalize IS weights?}

\item{wq}{Weights for contrast function}
}
\description{
...?
}
