% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{evaluate_improvement}
\alias{evaluate_improvement}
\title{Evaluate if there is sufficient improvement}
\usage{
evaluate_improvement(
  f_old,
  f_new,
  x_new,
  x_old,
  delta,
  deltaMax,
  eta,
  predicted_reduction
)
}
\arguments{
\item{f_old}{Value of function at evaluation location}

\item{f_new}{Value of function at evaluation location}

\item{x_new}{New parameter values}

\item{x_old}{Old parameter values}

\item{delta}{Old trust region radius}

\item{deltaMax}{Maximal trust region radius}

\item{eta}{Trust region parameter}

\item{predicted_reduction}{Reduction in objective with the model function}
}
\description{
evaluate improvement
}
