% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_SR1.R
\name{evaluate_improvement_SR1}
\alias{evaluate_improvement_SR1}
\title{Evaluate if there is sufficient improvement}
\usage{
evaluate_improvement_SR1(
  f_old,
  params_new,
  params_old,
  params_sim,
  predicted_reduction,
  simStepRes,
  K_hat,
  rho_hat,
  wq,
  normalized,
  delta_hat,
  eta
)
}
\arguments{
\item{f_old}{Value of function at evaluation location}

\item{params_new}{Parameter values at which to approximate the partial derivative}

\item{params_old}{Parameter values of the previous iteration}

\item{params_sim}{Parameter values used to simulate the ensemble in \code{simStepRes}}

\item{predicted_reduction}{Step for derivative calculation}

\item{simStepRes}{Result from the simulation step, see \code{\link{simulation_step}}. It must be the ensemble
simulated at \code{params_sim}, since that is the density the importance sampling weights
are taken against.}

\item{K_hat}{Estimated K-function for data}

\item{rho_hat}{Estimated intensity for data}

\item{wq}{Weights for contrast function}

\item{normalized}{Normalize IS weights?}

\item{delta_hat}{Old trust region radius}

\item{eta}{Trust region parameter}
}
\description{
evaluate improvement
}
