% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{find_alpha_range}
\alias{find_alpha_range}
\title{Function to determine allowable range for alpha in line search}
\usage{
find_alpha_range(x_k, x_0, p_k, delta_hat)
}
\arguments{
\item{x_k}{Previous iterate, origin of line search}

\item{x_0}{Midpoint of trust region}

\item{p_k}{Direction of line search}

\item{delta_hat}{Radius of trust region}
}
\description{
We want to do a line search satisfying ||x_k+alpha*p_k - x_0||_2 <= delta_hat
This is satisfied if
}
