% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{intensity_est}
\alias{intensity_est}
\title{Function for estimating the intensity given a list of point patterns
The point patterns are assumed to come from a time series of point patterns}
\usage{
intensity_est(points_input, l, timescale)
}
\arguments{
\item{points_input}{List of point patterns used for estimation}

\item{l}{Side length of observation window}

\item{timescale}{How many indices apart do snapshots need to be in order to be independent?}
}
\description{
Function for estimating the intensity given a list of point patterns
The point patterns are assumed to come from a time series of point patterns
}
