% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{K_lambda_specific_r}
\alias{K_lambda_specific_r}
\title{Estimate K_lambda for a specific radius}
\usage{
K_lambda_specific_r(r, pattern)
}
\arguments{
\item{r}{radius}

\item{pattern}{Pattern to estimate for}
}
\description{
Description
}
