% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast.R
\name{K_theo.exp}
\alias{K_theo.exp}
\title{Estimate K-function with exponential kernel}
\usage{
K_theo.exp(r_vec, params, win)
}
\arguments{
\item{r_vec}{Vector with radii to use for contrast calculation}

\item{params}{Parameter vector(to be optimized)}

\item{win}{Simulation window}
}
\description{
To be written
}
