% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_BFGS.R
\name{minimum_contrast_estimation}
\alias{minimum_contrast_estimation}
\title{Complete estimation procedure using importance sampling + simulation + minimum contrast}
\usage{
minimum_contrast_estimation(
  params,
  par_free_index,
  par_fixed,
  epsilon,
  imp_samp_tolerance,
  delta_hat,
  tol = 10^-8,
  max.iter = 10000,
  simulate_point_pattern
)
}
\arguments{
\item{params}{Initial guess for parameters to estimate}

\item{par_free_index}{Indices of parameters to estimate}

\item{par_fixed}{Values of fixed parameters}

\item{epsilon}{Step for derivative calculation}

\item{imp_samp_tolerance}{Tolerance for difference in parameter value before resampling patternSim}

\item{delta_hat}{trust radius}

\item{tol}{Convergence tolerance}

\item{max.iter}{Maximum iterations}

\item{simulate_point_pattern}{Function for simulation of point pattern}
}
\description{
Minimum contrast estimation using trust region approach with simulation based estimation
of gradient and SR1 for Hessian approximation.
This method also uses importance sampling...
}
