% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mypcfplot.ggplot}
\alias{mypcfplot.ggplot}
\title{Plot pcf using ggplot2}
\usage{
mypcfplot.ggplot(pcf_input)
}
\arguments{
\item{pcf_input}{Pcf function to be plotted}
}
\description{
Plot a pcf with ggplot, ignoring the first r-value in case this is NA
}
