% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast.R
\name{objective_weighted_contrast}
\alias{objective_weighted_contrast}
\title{Objective function for minimum contrast}
\usage{
objective_weighted_contrast(
  params,
  rho_hat,
  K_hat,
  params_fixed,
  free_params_indices = NULL,
  wpq = c(1, 1, 2, 1, 1/4, 2),
  ns_kernel = "var_gamma",
  nSims,
  win,
  union_est = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{params}{Parameter vector(to be optimized)}

\item{rho_hat}{Intensity estimated from data}

\item{K_hat}{K-function estimated from data}

\item{params_fixed}{Values for the fixed parameters}

\item{free_params_indices}{Which indices for the parameter vector are free?}

\item{wpq}{Vector of weights and powers (w_rho, q_rho, p_rho, w_K, q_K, p_K)}

\item{ns_kernel}{String indicating which kernel to use for the Neymann-Scott process, "var_gamma" or "thomas"}

\item{nSims}{How many simulations to estimate K_hat}

\item{win}{Simulation window}

\item{union_est}{Set to TRUE in order to use union approach for estimation of K-function}

\item{parallel}{Set to TRUE to use parallel computing when simulating point patterns for estimation}
}
\description{
Calculate contrast/objective function given params and estimates for the intensity and the K-function from data
}
