% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast.R
\name{objective_weighted_contrast_vargamma}
\alias{objective_weighted_contrast_vargamma}
\title{Objective function for minimum contrast}
\usage{
objective_weighted_contrast_vargamma(
  params,
  rho_hat,
  K_hat,
  repulsionRange_hat,
  w_rho = 1,
  q_rho = 1,
  p_rho = 2,
  w_K = 1,
  q_K = 1/4,
  p_K = 2,
  nSims,
  win,
  union_est = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{params}{Parameter vector(to be optimized)}

\item{rho_hat}{Intensity estimated from data}

\item{K_hat}{K-function estimated from data}

\item{repulsionRange_hat}{Set to FALSE if the repulsion range parameter is in the params vector}

\item{w_rho}{Weight of intensity}

\item{q_rho}{Power to use inside parantheses in contrast for intensity}

\item{p_rho}{Power to use outside parantheses in contrast for intensity}

\item{w_K}{Weight of K-function}

\item{q_K}{Power to use inside parantheses in contrast for K-function}

\item{p_K}{Power to use outside parantheses in contrast for K-function}

\item{nSims}{How many simulations to estimate K_hat}

\item{win}{Simulation window}

\item{union_est}{Set to TRUE in order to use union approach for estimation of K-function}

\item{parallel}{Set to TRUE to use parallel computing when simulating point patterns for estimation}
}
\description{
Calculate contrast/objective function given params and estimates for the intensity and the K-function from data
}
