% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{pcf_est.unions}
\alias{pcf_est.unions}
\title{Function for estimating the pair correlation function given a list of point patterns
The point patterns are assumed to come from a time series of point patterns}
\usage{
pcf_est.unions(
  points_input,
  bw = NULL,
  l,
  spacing,
  rMax,
  dr,
  timescale,
  divisor = "r",
  zerocor = NULL
)
}
\arguments{
\item{points_input}{List of point patterns used for estimation. Each element is a
\code{data.frame} with columns \code{x} and \code{y} in observation window coordinates, that is in
\verb{[0, l]}, or a bare number for a snapshot holding no points.}

\item{bw}{Estimation bandwidth}

\item{l}{Side length of observation window}

\item{spacing}{How far apart are observations window in order to be independent?}

\item{rMax}{Max radius for which estimation is performed}

\item{dr}{How fine grid for estimation?}

\item{timescale}{How many indices apart do snapshots need to be in order to be independent?}

\item{divisor}{See documentation from the spatstat-function pcf.ppp}

\item{zerocor}{See pcf.ppp documentation from spatstat}
}
\description{
Function for estimating the pair correlation function given a list of point patterns
The point patterns are assumed to come from a time series of point patterns
}
