% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rVarGamma_matern_thinned}
\alias{rVarGamma_matern_thinned}
\title{Variance Gamma with Matern thinning}
\usage{
rVarGamma_matern_thinned(
  kappa,
  scale,
  mu,
  nu,
  repulsionRange,
  win,
  thinningType = 2
)
}
\arguments{
\item{kappa}{See rVarGamma in spatstat}

\item{scale}{See rVarGamma in spatstat}

\item{mu}{See rVarGamma in spatstat}

\item{nu}{See rVarGamma in spatstat}

\item{repulsionRange}{Range of hard-core repulsion}

\item{win}{Simulation window to be used}

\item{thinningType}{Type of Matern thinning, 1, 2 or 3, see \code{\link{matern.thinning}}.}
}
\description{
Simulates a Variance Gamma SNCP and applies a Matern thinning of type \code{thinningType} to it
}
