% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turbulence}
\alias{turbulence}
\title{Coordinates of scars and dimples from simulated data}
\format{
\subsection{\code{turbulence}}{

List with the following items:
\describe{
\item{scars}{List of scars coordinates for each timestep}
\item{dimples}{List of dimple coordinates for each timestep}
\item{lengthscale, timescale}{turbulent scales}
\item{sidelength}{Sidelength of observation window}
\item{re, we}{Reynolds and Weber number of data}
...
}
}
}
\usage{
turbulence
}
\description{
Coordinates of scars and dimples from simulated data
}
\keyword{datasets}
