% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{contrast_function}
\alias{contrast_function}
\title{Calculate contrast function}
\usage{
contrast_function(rho_par, rho_hat, K_par, K_hat, wqp = c(100, 1/4, 2))
}
\arguments{
\item{rho_par}{Intensity for model and parameters}

\item{rho_hat}{Estimated intensity for data}

\item{K_par}{K-function for model and paramerters}

\item{K_hat}{Estimated K-function for data}

\item{wqp}{Weights for contrast function}
}
\description{
Take the estimated K-function and intensity,
together with the K-function and intensity for the parameters,
and calculates the contrast fucntion.
}
