% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_BFGS.R
\name{contrast_gradient_importance_sampling}
\alias{contrast_gradient_importance_sampling}
\title{Calculate partial derivative for the k'th parameter}
\usage{
contrast_gradient_importance_sampling(
  params,
  par_free_index,
  par_fixed,
  k,
  patternSim,
  epsilon,
  f_val
)
}
\arguments{
\item{params}{Parameter values at which to aproximate the partial derivative}

\item{par_free_index}{Indices of parameters to estimate}

\item{par_fixed}{Values of fixed parameters}

\item{k}{Index of direction to estimate derivative in}

\item{patternSim}{simulated point pattern to use}

\item{epsilon}{Step for derivative calculation}

\item{f_val}{Value of function at evaluation location}
}
\description{
Calculate partial derivative for the k'th parameter
}
