% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_SR1.R
\name{contrast_gradient_importance_sampling}
\alias{contrast_gradient_importance_sampling}
\title{Estimate gradient at some parameter value using importance sampling}
\usage{
contrast_gradient_importance_sampling(
  params,
  params_sim,
  par_free_index,
  epsilon,
  patternSim,
  rho_baseline,
  K_lambda_baseline,
  normalized,
  rho_hat,
  K_hat,
  wq,
  f_val
)
}
\arguments{
\item{params}{Parameter values at which to aproximate the partial derivative}

\item{params_sim}{Parameter values used in simulating patternSim}

\item{par_free_index}{Indices of parameters to estimate}

\item{epsilon}{Step for derivative calculation}

\item{patternSim}{simulated point pattern to use}

\item{rho_baseline}{rho_baseline}

\item{K_lambda_baseline}{K_lambda_baseline}

\item{normalized}{Normalize IS weights?}

\item{rho_hat}{Estimated intensity for data}

\item{K_hat}{Estimated K-function for data}

\item{wq}{Weights for contrast function}

\item{f_val}{Value of function at evaluation location}
}
\description{
Calculate partial derivative for the k'th parameter
}
