% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_evaluation.R
\name{contrast_is}
\alias{contrast_is}
\title{Calculate contrast function}
\usage{
contrast_is(
  patternSim,
  params_0,
  params_new,
  rho_hat,
  K_hat,
  rho_baseline,
  K_lambda_baseline,
  normalized = FALSE,
  wq = c(1000, 1/4)
)
}
\arguments{
\item{patternSim}{Simulated pattern}

\item{params_0}{Parameters used in simulation of the point pattern}

\item{params_new}{Target parameters}

\item{rho_hat}{Estimated intensity for data}

\item{K_hat}{Estimated K-function for data}

\item{rho_baseline}{rho_baseline}

\item{K_lambda_baseline}{K_lambda_baseline}

\item{normalized}{Normalize IS weights?}

\item{wq}{Weights for contrast function}
}
\description{
Take the estimated K-function and intensity,
together with the K-function and intensity for the parameters,
and calculates the contrast fucntion.
}
