% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{estimate_K_lambda_baseline}
\alias{estimate_K_lambda_baseline}
\title{Estimate the "unnormalized" K-function directly for a given point pattern}
\usage{
estimate_K_lambda_baseline(pattern, r_vec)
}
\arguments{
\item{pattern}{The point pattern to estimate for}

\item{r_vec}{Vector of radii to use}
}
\description{
Estimate the "unnormalized" K-function directly for a given point pattern, this function does only the "border" correction.
}
