% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{estimate_rho_baseline}
\alias{estimate_rho_baseline}
\title{Estimate the intensity directly for a given point pattern}
\usage{
estimate_rho_baseline(pattern)
}
\arguments{
\item{pattern}{The point pattern to estimate for}
}
\description{
Estimate the intensity directly for a given point pattern.
This estimate assumes stationarity.
}
