% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{evaluate_improvement}
\alias{evaluate_improvement}
\title{Evaluate if there is sufficient improvement}
\usage{
evaluate_improvement(
  f_old,
  params_new,
  params_old,
  params_sim,
  predicted_reduction,
  patternSim,
  K_hat,
  rho_hat,
  wqp,
  rho_baseline,
  K_lambda_baseline,
  normalized,
  delta_hat,
  eta
)
}
\arguments{
\item{f_old}{Value of function at evaluation location}

\item{params_new}{Parameter values at which to approximate the partial derivative}

\item{params_old}{Parameter values of the previous iteration}

\item{params_sim}{Parameter values used to simulate  patternSim}

\item{predicted_reduction}{Step for derivative calculation}

\item{patternSim}{simulated point pattern to use}

\item{K_hat}{Estimated K-function for data}

\item{rho_hat}{Estimated intensity for data}

\item{wqp}{Weights for contrast function}

\item{rho_baseline}{rho_baseline}

\item{K_lambda_baseline}{K_lambda_baseline}

\item{normalized}{Normalize IS weights?}

\item{delta_hat}{Old trust region radius}

\item{eta}{Trust region parameter}
}
\description{
evaluate improvement
}
