% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust_step.R
\name{find_alpha_range}
\alias{find_alpha_range}
\title{Function to determine allowable range for alpha in line search}
\usage{
find_alpha_range(x_k, x_0, p_k, delta)
}
\arguments{
\item{x_k}{Previous iterate, origin of line search}

\item{x_0}{Midpoint of trust region}

\item{p_k}{Direction of line search}

\item{delta}{Radius of trust region}
}
\description{
We want to do a line search satisfying ||x_k+alpha*p_k - x_0||_2 <= delta

This function solves the simple quadratic equation needed to find the range of
alpha values that satisfies thiat requirement.
}
