% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{find_min_dist}
\alias{find_min_dist}
\title{Minimum inter-point distance for list of points}
\usage{
find_min_dist(points_input, timescale = 1, l)
}
\arguments{
\item{points_input}{List of point patterns used for estimation}

\item{timescale}{How many indices apart do snapshots need to be in order to be independent?}

\item{l}{Side length of observation window}
}
\description{
Minimum inter-point distance for list of points
}
