% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{importance_sampling_weigths}
\alias{importance_sampling_weigths}
\title{Use importance sampling to estimate the K-function}
\usage{
importance_sampling_weigths(
  kappa,
  mu,
  omega,
  kappa_0 = NULL,
  mu_0 = NULL,
  omega_0 = NULL,
  patternSim,
  log_f_kappa_0 = NULL,
  log_fCond_theta_0 = NULL,
  parallel = FALSE
)
}
\arguments{
\item{kappa}{Kappa parameter}

\item{mu}{mu parameter}

\item{omega}{omega parameter}

\item{kappa_0}{Kappa parameter}

\item{mu_0}{mu parameter}

\item{omega_0}{omega parameter}

\item{patternSim}{List of point patterns to use for estimation}

\item{log_f_kappa_0}{Parent null density}

\item{log_fCond_theta_0}{Daughter null density}

\item{parallel}{Set to TRUE to use parallel computing}
}
\description{
Use importance sampling to estimate the Ripleys' K-function for parameter values
}
