% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_core_functions.R
\name{K_importance_sampling}
\alias{K_importance_sampling}
\title{Use importance sampling to estimate the K-function}
\usage{
K_importance_sampling(
  w_is,
  K_lambda_baseline,
  rho_baseline,
  r_vec = NULL,
  patternSim = NULL,
  normalized = FALSE
)
}
\arguments{
\item{w_is}{Importance sampling weights}

\item{K_lambda_baseline}{K_lambda_baseline}

\item{rho_baseline}{rho_baseline}

\item{r_vec}{r_vec}

\item{patternSim}{List of point patterns to use for estimation}

\item{normalized}{Set to TRUE if you want to normalize the importance sampling ratios}
}
\description{
Use importance sampling to estimate the Ripleys' K-function for parameter values
}
