% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{matern.thinning}
\alias{matern.thinning}
\title{Matern type II thinning}
\usage{
matern.thinning(initialPattern, repulsionRange, xrange, yrange)
}
\arguments{
\item{initialPattern}{DataFrame with the unthinned point pattern}

\item{repulsionRange}{Range of hard-core repulsion}

\item{xrange}{vector with max and min of x-axis in observation window}

\item{yrange}{vector with max and min of y-axis in observation window}
}
\description{
Perform a Matern type II thinning on a point pattern
}
