% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{min_contrast_trust_region}
\alias{min_contrast_trust_region}
\title{Complete estimation procedure using importance sampling + simulation + minimum contrast}
\usage{
min_contrast_trust_region(
  params,
  parFreeIndex,
  repRange,
  rho_hat,
  K_hat,
  xlims,
  ylims,
  nSims,
  deltaInit,
  eta,
  deltaMax,
  deltaMin = 1e-04,
  wq = c(1000, 1/4),
  normalized = FALSE,
  countPrefitLoops = 2,
  max.iter.prefit = 0,
  tol = 10^-8,
  max.iter = 1000,
  printProgress = FALSE
)
}
\arguments{
\item{params}{Initial guess for parameters to estimate}

\item{parFreeIndex}{Indices of parameters to estimate}

\item{repRange}{Repulsion range}

\item{rho_hat}{Estimated intensity for data}

\item{K_hat}{Estimated K-function for data}

\item{xlims}{Simulation window x limits}

\item{ylims}{Simulation window y limits}

\item{nSims}{Number of simulations}

\item{deltaInit}{Initial trust radius}

\item{eta}{Trust region parameter}

\item{deltaMax}{Greatest trust region radius allowed}

\item{deltaMin}{Convergence tolerance for trust region radius}

\item{wq}{Weights for contrast function}

\item{normalized}{Normalize IS weights?}

\item{max.iter.prefit}{How many iterations to do at most for the pre-fitting}

\item{tol}{Convergence tolerance}

\item{max.iter}{Maximum iterations}

\item{printProgress}{Set to TRUE to get updates on progress while running}
}
\description{
Minimum contrast estimation using trust region approach with simulation based estimation
of gradient and SR1 for Hessian approximation.
This method also uses importance sampling...
}
