% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{parent_log_density}
\alias{parent_log_density}
\title{Calculate Poisson density of a parent process}
\usage{
parent_log_density(pattern = NULL, kappa, parent = NULL, B_area = NULL)
}
\arguments{
\item{pattern}{A list containing elements "parent" and "B_area" as described above}

\item{kappa}{Positive scalar. Intensity of parent points.}

\item{parent}{A \code{data.frame} with columns \code{x} and \code{y} giving parent locations,
typically restricted to an enlarged window \eqn{B \supset W}.}

\item{B_area}{Area of the enlarged window \eqn{B \supset W}.}
}
\description{
Computes the log-density of a daughter point pattern \eqn{C} as a homogeneous
Poisson Point pattern restricted to the window \eqn{B}, with intensity \eqn{\kappa}.

Specifically, the density is given by
\deqn{f(C\cup B|\kappa)=\exp\left(|B|(1-\kappa)\right)\kappa^{|C\cap B|}}
}
