% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{result_verification}
\alias{result_verification}
\title{Compare the results from optimization with the target K-function and intensity}
\usage{
result_verification(
  rho_hat,
  K_hat,
  param_estim,
  repRange_estim,
  q = 1/4,
  nSims = 1000,
  xrange = c(0, 10),
  yrange = c(0, 10)
)
}
\arguments{
\item{rho_hat}{Intensity estimated from the data}

\item{K_hat}{K-function estimated from the data}

\item{param_estim}{The parameters estimated from the minimum contrast estimation}

\item{repRange_estim}{Estimated repulsion range}

\item{q}{Power to apply to K-function}

\item{nSims}{numer of simulations}

\item{xrange}{xrange for simulation window}

\item{yrange}{yrange for simulation window}
}
\description{
Compare the results from optimization with the target K-function and intensity
}
