% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling_core_functions.R
\name{rho_importance_sampling}
\alias{rho_importance_sampling}
\title{Use importance sampling to estimate the (stationary) intensity}
\usage{
rho_importance_sampling(
  w_is,
  rho_baseline,
  patternSim = NULL,
  normalized = FALSE
)
}
\arguments{
\item{w_is}{Importance sampling weights}

\item{rho_baseline}{rho_baseline}

\item{patternSim}{List of point patterns to use for estimation}

\item{normalized}{Set to TRUE if you want to normalize the importance sampling ratios}
}
\description{
Using a thinned Thomas process simulated with parameters kappa_0, mu_0, and omega_0,
we use importance sampling to estimate
}
