% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rThomas_matern_thinned}
\alias{rThomas_matern_thinned}
\title{Thomas process with Matern II thinning}
\usage{
rThomas_matern_thinned(
  kappa,
  scale,
  mu,
  repulsionRange,
  xlims,
  ylims,
  saveparents = FALSE
)
}
\arguments{
\item{kappa}{See rThomas in spatstat}

\item{scale}{See rThomas in spatstat}

\item{mu}{See rThomas in spatstat}

\item{repulsionRange}{Range of hard-core repulsion}

\item{xlims}{xlim}

\item{ylims}{ylim}

\item{saveparents}{Logical value indicating whether to save the locations of the parent points as an attribute.}
}
\description{
Simulates a Variance Gamma SNCP and applies a Matern II thinning to it
}
