% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rVarGamma_matern_thinned}
\alias{rVarGamma_matern_thinned}
\title{Variance Gamma with Matern II thinning}
\usage{
rVarGamma_matern_thinned(kappa, scale, mu, nu, repulsionRange, win)
}
\arguments{
\item{kappa}{See rVarGamma in spatstat}

\item{scale}{See rVarGamma in spatstat}

\item{mu}{See rVarGamma in spatstat}

\item{nu}{See rVarGamma in spatstat}

\item{repulsionRange}{Range of hard-core repulsion}

\item{win}{Simulation window to be used}
}
\description{
Simulates a Variance Gamma SNCP and applies a Matern II thinning to it
}
