% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{simulation_step}
\alias{simulation_step}
\title{Simulation step in trust region optimization}
\usage{
simulation_step(
  nSims,
  params,
  repRange,
  xlims,
  ylims,
  K_hat,
  printProgress = FALSE
)
}
\description{
This function performs the simulation step in our optimizer.
}
