% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_SR1.R
\name{SR1_update}
\alias{SR1_update}
\title{Update the secant matrix with the SR1 update rule}
\usage{
SR1_update(Bk, sk, r, yk)
}
\arguments{
\item{Bk}{N&W}

\item{sk}{N&W}

\item{r}{N&W}

\item{yk}{N&W}
}
\description{
See 6.2 in Noccedal and Wright
}
