% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{trust_line_steps}
\alias{trust_line_steps}
\title{Line minimization using interpolation with maximum reach}
\usage{
trust_line_steps(
  p_k,
  x_k,
  x_0,
  delta_hat,
  trust_function,
  bisection_count = 11
)
}
\arguments{
\item{p_k}{Search direction}

\item{x_k}{Search starting point}

\item{x_0}{Center of trust region}

\item{delta_hat}{Radius of trust region}

\item{trust_function}{Trust function}

\item{bisection_count}{Number of bisections}
}
\description{
Line optimization for the trust steps
}
