% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust_step.R
\name{trust_step}
\alias{trust_step}
\title{Do one iteration within trust region}
\usage{
trust_step(x_0, delta, trust_function)
}
\arguments{
\item{x_0}{Parameter values used to simulate patternSim}

\item{delta}{Trust region radius}

\item{trust_function}{Trust function}
}
\description{
Iterate from params_sim using importance sampling on the realizations patternSim, simulated from params_sim.
The iteration is restricted to the ball centered on the parameters used for simulation,
with radius equal to the trust region radius.
}
