% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_contrast_trust_region.R
\name{trust_step}
\alias{trust_step}
\title{Do one iteration within trust region}
\usage{
trust_step(
  f_old,
  patternSim,
  K_lambda_baseline,
  rho_baseline,
  params_sim,
  delta_hat
)
}
\description{
Do one
}
